fixed problem of definition of tagmap
