+++ git_vuln_finder/__init__.py
@@ -1,6 +1,6 @@
 
-from git_vuln_finder.finder import build_pattern
-from git_vuln_finder.finder import get_patterns
-from git_vuln_finder.finder import find_vuln
-from git_vuln_finder.finder import summary
-from git_vuln_finder.finder import extract_cve
+from git_vuln_finder.pattern import build_pattern
+from git_vuln_finder.pattern import get_patterns
+from git_vuln_finder.vulnerability import find_vuln
+from git_vuln_finder.vulnerability import summary
+from git_vuln_finder.vulnerability import extract_cve
